Modernize utils.py: snake_case, type hints, move dead code to example script
